remove api dir
